+++ src/os2datascanner/engine2/model/derived/utilities/ghostscript.py
@@ -2,6 +2,7 @@
 Contains utilities for compressing PDF files using GhostScript.
 """
 import shlex
+import structlog
 
 from tempfile import TemporaryDirectory
 
@@ -11,20 +12,43 @@ from .... import settings as engine2_settings
 GS = engine2_settings.ghostscript
 
 
+logger = structlog.get_logger("engine2")
+
+
 def gs_convert(path):
     '''Convert to a compressed form using GhostScript (gs).'''
     with TemporaryDirectory() as outputdir:
+        match GS["_base_arguments"]:
+            case list() as bl:
+                base_arguments = bl
+            case str() as bs:
+                logger.warning(
+                        "deprecated type for setting"
+                        " 'ghostscript._base_arguments';"
+                        " use a list instead")
+                base_arguments = shlex.split(bs)
+            case _:
+                raise TypeError
+
+        match GS["extra_args"]:
+            case list() as el:
+                extra_args = el
+            case str() as es:
+                logger.warning(
+                        "deprecated type for setting"
+                        " 'ghostscript.extra_args';"
+                        " use a list instead")
+                base_arguments = shlex.split(es)
+            case _:
+                raise TypeError
+
         converted_path = "{0}/gs-temporary.pdf".format(outputdir)
         command = ["gs", "-q",
-                   *shlex.split(GS["_base_arguments"]),
+                   *base_arguments,
                    f"-dPDFSETTINGS={GS['pdf_profile']}",
-                   "-sOutputFile={0}".format(converted_path)]
-
-        extra_args = shlex.split(GS["extra_args"])
-        if extra_args:
-            command.extend(extra_args)
-
-        command.append(path)
+                   "-sOutputFile={0}".format(converted_path),
+                   *extra_args,
+                   path]
 
         run_custom(command,
                    timeout=GS["timeout"],